backend/pyproject.toml
[tool.poetry]
name = "berlin-guide"
version = "0.1.0"
description = "The backend of the Berlin Guide for Romanians"
authors = ["Theo Costea <theo.costea@gmail.com>"]
license = "Apache License"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.7"
djangorestframework = "^3.12.4"
psycopg2-binary = "^2.8.6"
djangorestframework-gis = "^0.17"
django-cors-headers = "^3.7.0"
httpx = "^0.17.1"
geocoder = "^1.38.1"
django-filter = "^2.4.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pytest = "^6.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
